test_mweb/requirements.txt
Appium-Python-Client     5.3.0
pytest                   8.4.1
pytest-xdist             3.8.0
requests                 2.32.4
requests-oauthlib        2.0.0
selenium                 4.33.0
pytest-order             1.3.0
pytest-html              4.2.0
selenium                 4.33.0

